Created django rest api with user and snapcapsule model
Created django app called api. API holds the scheme for snapcapsule in api/models.py. I am also using the django preset user model. The next commit will be to deploy this app heroku. Also, it would be nice to add foriegn key to snapcapusle for the username of the user.