Authentication setup up with jwt-token
The login feature is working correctly. The sigir will be a commit to make the signup feature work correctly. I am using library to decode the jwt token and fetch command to send that data to the backend. In the backend, I install jwt and added urls to refresh and get tokens