Setup backend support for users to post capsules
The delete function in class Snapcapsule of models.py was overriden,
so that deleting the Snapcapsule object will delete the file. This code
will probably not work in a online repo. A commit will probably update
this function in the future.

In Serializers.py, I added an update function, so Snapcapsules can be
posted to a User. This is needed in order to store the hyperLinkUrl of
the capsules in User data. There will be a commit in the future to not
allow Capsule to create without relation to user, so UserSerializer
object will have a link to the users capsules.

The parser classes were deleted from views.py because they were not
needed to parse form-data. Django Rest Framework or Postman offers
defualt headers to parse form-data.