Implement display capsule feature
In the backend, I added the username to serialized data, so the frontend
will be able to display the username of the capsule. I added the
media_root to the urls, so the frontend can request the images saved
in the backend through urls instead of the directory path.

In the frontend, Axios makes request to the backend for capsules. A
commit in the feature will allow for pagination. In order to make sure
that data is deleted properly, I add clear function to snapcapsule
class.

Error boundary were also added to frontend for easier debugging. Imports
were structured in alphabetical order for a better read.