Deployed frontend to AWS s3 server
postcss.config.js was added in order to make the postcss module load
correctly. A command was added to package.json to build and deploy
changes to the AWS s3 bucket. Dist contains the current version of the
site at http://snaphireme.s3-website.us-west-1.amazonaws.com/

The backend setitngs.py was change to a directory for a cleaner
deployment. Base.py will serve as the shared settings between
development and production. Also, a new secret was generated and stored
in file named in the gitignore. Sorry hackers.