Created feature to POST SnapCapsules from frontend
The node modules have been added to make the repository up to date

Hero.js image were change to imports for site. The login form and
snapcapsules form recieved form validation, so the backend does not have
to deal impromptu submit. Validation will be broken up in future
commits.

Snapcapsules are being posted to the backend, but UserSerializer still
does not have any reference to their snapcapsules. The solution to this
will be in the next commit.